pyproject.toml
[tool.poetry]
name = "in-trust"
version = "0.1.0"
description = ""
authors = ["knowlegbox <kieriekhenoc@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.1"
psycopg2 = "^2.9.10"
firebase-admin = "^6.8.0"
django-cors-headers = "^4.7.0"
graphene-django-query-optimizer = "^0.10.11"
graphene-django = "^3.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
